requirements.txt
# food detector requirements
opencv-python
ultralytics

# recipe generator requirements 
huggingface_hub>=0.20.0
python-dotenv>=1.0.0
